% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countaacateg.R
\name{count.aa.categ}
\alias{count.aa.categ}
\title{Count Amino Acid Categories}
\usage{
count.aa.categ(file)
}
\arguments{
\item{file}{A file connection. This should be a fasta-format text file of the desired protein.}
}
\value{
A list of 5 integers, each representing the total number of amino acids within a broad category that the protein contains.
}
\description{
This function takes a fasta file and creates a list that has the total number of amino acids in each category.
}
\details{
This function divides the 20 amino acids into 5 categories: Hydrophobic, Hydrophilic, Ambiguous, Acidic, and Basic.  The Ambiguous category contains cysteine and tyrosine; different sources disagree on whether these amino acids are hydrophobic or hydrophilic.
}
