% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countaatype.R
\name{count.aa.type}
\alias{count.aa.type}
\title{Count Amino Acid Types}
\usage{
count.aa.type(file)
}
\arguments{
\item{file}{A file connection. This should be a fasta-format text file of the desired protein.}
}
\value{
A list of 20 integers, each representing the total number of an amino acid type in the protein.
}
\description{
This function creates a list that contains the number of each type of amino acid that can be found in a provided fasta file.
}
\details{
The returned list contains the amino acid one-letter codes as the keys, and the matching integers are the total number of that amino acid.  For example, the element "G=16" indicates that the protein has 16 glycines.
}
