% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{get.seq}
\alias{get.seq}
\title{Get Sequence}
\usage{
get.seq(file)
}
\arguments{
\item{file}{A file connection.  This should be a fasta-format text file of the desired protein.}
}
\value{
A string of the amino acid one-letter codes, taken from the file argument.
}
\description{
Using a fasta file, this function returns the amino acid sequence as a single string.
}
\details{
This function removes the first line of the fasta file, which is the header.  It concatenates the remaining lines into one string to create an amino acid sequence.
}
