% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphaacateg.R
\name{graph.aa.categ}
\alias{graph.aa.categ}
\title{Graph Amino Acid Categories}
\usage{
graph.aa.categ(file)
}
\arguments{
\item{file}{A file connection. This should be a fasta-format text file of the desired protein.}
}
\value{
A ggplot object.
}
\description{
This function creates a bar graph using ggplot2.
The graph displays the amount of amino acids within each broad category that can be found in a provided fasta file.
}
\details{
The x-axis of the graph contains the broad categories of amino acids, and the y-axis, or bar height, displays the amount of amino acids in each category in the protein.
The Ambiguous category contains cysteine and tyrosine; different sources disagree on whether these amino acids are hydrophobic or hydrophilic.
}
