% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-aa.R
\name{graph.aa}
\alias{graph.aa}
\title{Graph Amino Acids}
\usage{
graph.aa(file)
}
\arguments{
\item{file}{A file connection. This should be a fasta-format text file of the desired protein.}
}
\value{
A ggplot object.
}
\description{
This function creates a bar graph using ggplot2.
The graph displays the amount of each of the 20 amino acids that can be found in a provided fasta file.
}
\details{
The x-axis of the graph contains the amino acid one-letter codes, and the y-axis, or bar height, displays the amount of each amino acid in the protein.
}
