% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-length.R
\name{protein.length}
\alias{protein.length}
\title{Protein Length}
\usage{
protein.length(file)
}
\arguments{
\item{file}{A file connection. This should be a fasta-format text file of the desired protein.}
}
\value{
An integer representing the number of amino acids in the protein.
}
\description{
This function determines the length of a protein in a given fasta file.
}
